Refactor again before code review
