+++ frontend/src/components/shared/form-elements/textField.tsx
@@ -31,7 +31,7 @@ export const TextField: FC<TextFieldProps> = ({
         onChange={onChange}
         placeholder={placeholder}
         required={required}
-        rows={5}
+        rows={3}
         className="whitespace-pre-wrap w-full px-4 py-2 mt-1 border rounded-md focus:ring-2 focus:ring-blue-500 focus:outline-none focus:border-transparent"
       />
     </div>
